pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "unified-incident-env"
version = "1.0.0"
description = "Unified OpenEnv benchmark for incident response with causally linked WebSec remediation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "websockets>=12.0",
    "rich>=13.0.0",
    "matplotlib>=3.9.0",
    "numpy>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0"
]

[project.scripts]
server = "server.app:main"
baseline = "unified_incident_env.scripts.baseline_agent:main"
walkthrough = "unified_incident_env.scripts.walkthrough:main"
trainer-run-episode = "unified_incident_env.trainer.run_episode:main"
trainer-build-dataset = "unified_incident_env.trainer.build_sft_dataset:main"
trainer-eval-models = "unified_incident_env.trainer.eval_models:main"
trainer-build-datasets = "unified_incident_env.trainer.build_datasets:main"
trainer-update-model = "unified_incident_env.trainer.update_model:main"
trainer-run-session = "unified_incident_env.trainer.run_session:main"
trainer-train-external = "unified_incident_env.trainer.train_external:main"

[tool.hatch.build.targets.wheel]
packages = ["unified_incident_env", "server"]
